# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "客服系统api `\n客服后端：`",
        "title": "客服系统API文档`",
        "contact": {},
        "version": "1.0`"
    },
    "paths": {
...
        "/service/auth/login": {
            "post": {
                "tags": [
                    "客服系统"
                ],
                "summary": "客服登录接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录账号",
                        "name": "username",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/service/config": {
            "post": {
                "tags": [
                    "客服系统"
                ],
                "summary": "公共配置-获取系统配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/service/count": {
            "post": {
                "tags": [
                    "客服系统"
                ],
                "summary": "统计",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/service/info": {
            "post": {
                "description": "service_manager_id 账号ID\nServiceId 客服ID\nFileName 名称\nHead 头像\nUsername 账号\nCode code码 暂时没有用\nDomain 微信入口码\nHost 绑定的域名\nWeb 微信入口链接\nTimeOut 过期时间\nCreateTime 创建时间",
                "tags": [
                    "客服系统"
                ],
                "summary": "获取账号信息详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/service/notice_setting/info": {
            "post": {
                "tags": [
                    "客服系统"
                ],
                "summary": "公告-详细信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/service/notice_setting/update": {
            "post": {
                "tags": [
                    "客服系统"
                ],
                "summary": "公告-修改",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "启用状态",
                        "name": "is_show",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "图片",
                        "name": "image",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "内容",
                        "name": "text",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/service/reset_qrcode": {
            "post": {
                "tags": [
                    "客服系统"
                ],
                "summary": "重置二维码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/service/service_manager/message/list": {
            "post": {
                "tags": [
                    "客服系统"
                ],
                "summary": "快捷回复管理-获取消息列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "指定页",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分页数量",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/service/setting": {
            "post": {
                "tags": [
                    "客服系统"
                ],
                "summary": "公共配置-公告列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "类型",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "搜索值",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "指定页",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分页数量",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/service/update": {
            "post": {
                "tags": [
                    "客服系统"
                ],
                "summary": "修改客服和昵称",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "客服名称",
                        "name": "head",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "客服头像",
                        "name": "head",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "二维码背景",
                        "name": "code_background",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "二维码颜色",
                        "name": "code_color",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/service/users": {
            "post": {
                "tags": [
                    "客服系统"
                ],
                "summary": "用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "搜索值",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "指定页",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分页数量",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    }
}